requirements.txt
pandas
openpyxl 
requests
selenium
exceptiongroup
openai
python-dotenv